% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.tables.R
\name{compare.tables}
\alias{compare.tables}
\title{Compare by ID}
\usage{
compare.tables(table1, table2, id)
}
\arguments{
\item{table1}{First data frame}

\item{table2}{Second data frame}

\item{id}{Name of the ID column}
}
\value{
A data frame with discrepancies
}
\description{
this function compares to tables by ID column.
}
